Reopen AJOSF-72 onboarding should support ConnectionResult signal

+++ cpp/samples/OnboardingClientSample/OnboardingSignalListenerImpl.cc
@@ -28,7 +28,7 @@ OnboardingSignalListenerImpl::~OnboardingSignalListenerImpl()
 }
 
 
-void OnboardingSignalListenerImpl::ConnectionResultSignalReceived(int connectionResultCode, const qcc::String& connectionResultMessage)
+void OnboardingSignalListenerImpl::ConnectionResultSignalReceived(short connectionResultCode, const qcc::String& connectionResultMessage)
 {
     std::cout << "connectionResultCode=" << connectionResultCode << std::endl;
     std::cout << "connectionResultMessage=" << connectionResultMessage.c_str() << std::endl;
